backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
python-dotenv
argon2-cffi
PyJWT
Flask-Mail
gunicorn
pandas
openpyxl
resend
requests
